[TaskManager] Split filesystem API to seperate module
