+++ TaskManager/TaskManager-Core/build.gradle
@@ -12,6 +12,7 @@ compileJava {
 
 dependencies {
     implementation project(':Messenger:MessengerClient')
+    implementation project(':TaskManager:TaskManager-FileSystem')
 
     implementation 'org.zeroturnaround:zt-exec:1.12'
     implementation 'com.google.code.gson:gson:2.9.0'
@@ -22,6 +23,7 @@ dependencies {
 
 jar {
     dependsOn ':Messenger:MessengerClient:jar'
+    dependsOn ':TaskManager:TaskManager-FileSystem'
 
     manifest {
         attributes 'Main-Class': 'com.swrobotics.taskmanager.TaskManagerMain'
