+++ TaskManager/TaskManager-Core/src/main/java/com/swrobotics/taskmanager/FileTypeAdapter.java
