+++ TaskManager/TaskManager-Core/src/main/java/com/swrobotics/taskmanager/TaskOutputLogger.java
