+++ TaskManager/TaskManager-FileSystem/build.gradle
@@ -1,0 +1,19 @@
+plugins {
+    id 'java-library'
+}
+
+group 'com.swrobotics'
+version '2023'
+
+compileJava {
+    sourceCompatibility = '11'
+    targetCompatibility = '11'
+}
+
+dependencies {
+    api project(':Messenger:MessengerClient')
+}
+
+jar {
+    dependsOn ':Messenger:MessengerClient:jar'
+}
